OAuth2.0/requirements.txt
Flask==2.0.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
requests==2.27.0
SQLAlchemy==1.4.29